Refactor commands into modules
