+++ commands/cache.py
@@ -1,0 +1,192 @@
+import json
+import re
+from pathlib import Path
+from datetime import datetime
+
+from utils import debug_print, normalize_url
+
+CACHE_DIR = Path("migration_cache")
+CACHE_DIR.mkdir(exist_ok=True)
+
+
+def _cache_page_data(state, url, data):
+    domain = state.get_variable("DOMAIN")
+    row = state.get_variable("ROW")
+    kanban_id = state.get_variable("KANBAN_ID")
+    selector = state.get_variable("SELECTOR")
+    include_sidebar = state.get_variable("INCLUDE_SIDEBAR")
+    url = normalize_url(url)
+
+    if domain and row:
+        cache_filename = f"page_check_{domain}-{row}.json"
+    else:
+        sanitized_url = re.sub(r"[^\w\-_.]", "_", url)[:50]
+        cache_filename = f"page_check_{sanitized_url}.json"
+
+    cache_file = CACHE_DIR / cache_filename
+
+    cache_data = {
+        "metadata": {
+            "url": url,
+            "domain": domain,
+            "row": row,
+            "kanban_id": kanban_id,
+            "selector": selector,
+            "include_sidebar": include_sidebar,
+            "timestamp": datetime.now().isoformat(),
+            "cache_filename": cache_filename,
+        },
+        "page_data": data,
+    }
+
+    with open(cache_file, "w", encoding="utf-8") as f:
+        json.dump(cache_data, f, indent=2, ensure_ascii=False)
+
+    state.set_variable("CACHE_FILE", str(cache_file))
+    print(f"✅ Data cached to {cache_file}")
+
+
+def cache_page_data(state, url, data):
+    return _cache_page_data(state, url, data)
+
+
+def _load_cached_page_data(cache_file_path):
+    try:
+        with open(cache_file_path, "r", encoding="utf-8") as f:
+            cached_content = json.load(f)
+
+        if (
+            isinstance(cached_content, dict)
+            and "metadata" in cached_content
+            and "page_data" in cached_content
+        ):
+            return cached_content["metadata"], cached_content["page_data"]
+        else:
+            return {}, cached_content
+    except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
+        debug_print(f"Error loading cache file {cache_file_path}: {e}")
+        return {}, {}
+
+
+def _is_cache_valid_for_context(state, cache_file):
+    if not cache_file:
+        return False, "No cache file specified"
+    try:
+        metadata, _ = _load_cached_page_data(cache_file)
+        if not metadata:
+            return False, "Cache file contains no metadata"
+
+        current_url = state.get_variable("URL")
+        current_domain = state.get_variable("DOMAIN")
+        current_row = state.get_variable("ROW")
+        current_include_sidebar = state.get_variable("INCLUDE_SIDEBAR")
+
+        cached_url = metadata.get("url")
+        cached_domain = metadata.get("domain")
+        cached_row = metadata.get("row")
+        cached_include_sidebar = metadata.get("include_sidebar", False)
+
+        if current_url and cached_url:
+            url_matches = normalize_url(cached_url) == normalize_url(current_url)
+            if not url_matches:
+                return False, f"URL mismatch: cached={cached_url}, current={current_url}"
+
+        if current_domain and cached_domain and cached_domain != current_domain:
+            return False, f"Domain mismatch: cached={cached_domain}, current={current_domain}"
+
+        if current_row and cached_row and cached_row != current_row:
+            return False, f"Row mismatch: cached={cached_row}, current={current_row}"
+
+        sidebar_compatible = (not current_include_sidebar) or (
+            current_include_sidebar and cached_include_sidebar
+        )
+        if not sidebar_compatible:
+            return (
+                False,
+                f"Sidebar compatibility: need sidebar={current_include_sidebar}, cached sidebar={cached_include_sidebar}",
+            )
+        return True, "Cache is valid for current context"
+    except Exception as e:
+        debug_print(f"Error validating cache: {e}")
+        return False, f"Error validating cache: {e}"
+
+
+def _find_cache_file_for_domain_row(domain, row):
+    if not domain or not row:
+        return None
+    cache_filename = f"page_check_{domain}-{row}.json"
+    cache_file = CACHE_DIR / cache_filename
+    if cache_file.exists():
+        debug_print(f"Found cache file for {domain}-{row}: {cache_file}")
+        return str(cache_file)
+    debug_print(f"No cache file found for {domain}-{row}")
+    return None
+
+
+def _find_cache_file_for_url(url):
+    if not url:
+        return None
+    url = normalize_url(url)
+    for cache_file in CACHE_DIR.glob("page_check_*.json"):
+        try:
+            metadata, _ = _load_cached_page_data(cache_file)
+            cached_url = metadata.get("url")
+            if cached_url and normalize_url(cached_url) == url:
+                debug_print(f"Found cache file for URL {url}: {cache_file}")
+                return str(cache_file)
+        except Exception as e:
+            debug_print(f"Error checking cache file {cache_file}: {e}")
+            continue
+    debug_print(f"No cache file found for URL: {url}")
+    return None
+
+
+def _update_cache_file_state(state, url=None, domain=None, row=None):
+    current_cache_file = state.get_variable("CACHE_FILE")
+    search_url = url or state.get_variable("URL")
+    search_domain = domain or state.get_variable("DOMAIN")
+    search_row = row or state.get_variable("ROW")
+
+    found_cache_file = None
+    if search_domain and search_row:
+        found_cache_file = _find_cache_file_for_domain_row(search_domain, search_row)
+    if not found_cache_file and search_url:
+        found_cache_file = _find_cache_file_for_url(search_url)
+
+    if current_cache_file and found_cache_file != current_cache_file:
+        debug_print(
+            f"Current cache file {current_cache_file} doesn't match new context, unsetting"
+        )
+        state.set_variable("CACHE_FILE", "")
+        state.current_page_data = None
+
+    if found_cache_file:
+        state.set_variable("CACHE_FILE", found_cache_file)
+        try:
+            metadata, page_data = _load_cached_page_data(found_cache_file)
+            if page_data:
+                state.current_page_data = page_data
+                if metadata:
+                    if metadata.get("kanban_id"):
+                        state.set_variable("KANBAN_ID", metadata["kanban_id"])
+                    if metadata.get("url"):
+                        state.set_variable("URL", metadata["url"])
+                    if metadata.get("domain"):
+                        state.set_variable("DOMAIN", metadata["domain"])
+                    if metadata.get("row"):
+                        state.set_variable("ROW", metadata["row"])
+                    if metadata.get("selector"):
+                        state.set_variable("SELECTOR", metadata["selector"])
+                    if metadata.get("include_sidebar") is not None:
+                        state.set_variable("INCLUDE_SIDEBAR", str(metadata["include_sidebar"]).lower())
+                print(f"📋 Loaded cached data from {Path(found_cache_file).name}")
+            else:
+                debug_print(
+                    f"Cache file {found_cache_file} exists but contains no page data"
+                )
+        except Exception as e:
+            debug_print(f"Error loading cached page data from {found_cache_file}: {e}")
+    elif not found_cache_file and current_cache_file:
+        state.set_variable("CACHE_FILE", "")
+        state.current_page_data = None
+        debug_print("No matching cache file found, unset CACHE_FILE")
