+++ test_refactor.py
@@ -8,7 +8,7 @@ import os
 
 sys.path.insert(0, os.path.dirname(__file__))
 
-from commands.core import _is_cache_valid_for_context
+from commands.cache import _is_cache_valid_for_context
 from utils import normalize_url
